// frontend/school-management/src/app/core/models/enums/notification-type.enum.ts
export enum NotificationType {
  IMPORT_ERROR = 'import_error',
  IMPORT_FINISHED = 'import_finished',
  INVITATION_ACCEPTED = 'invitation_accepted',
  INVITATION_FAILED = 'invitation_failed',
  RELATION_DELETED = 'relation_deleted',
  PAYMENT_CONCEPT_STATUS_CHANGED = 'payment_concept_status_changed',
  PAYMENT_CONCEPT_CHANGED = 'payment_concept_changed',
  PROMOTION_COMPLETED = 'promotion_completed',
  PROMOTION_FAILED = 'promotion_failed',
}